requirements.txt
flask
requests
python-nmap
cryptography
pillow
numpy